shop_backend/requirements/base.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
django-filter>=23.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
Pillow==10.2.0
psycopg[binary]==3.1.19
sqlparse==0.5.5
celery==5.3.6
redis==5.0.3
